# Call RRDs::fetch() with a zero length time period